优化坦克随机移动

+++ src/main/java/group/li/pojo/Tank.java
@@ -99,25 +99,25 @@ public class Tank implements GetInfo {
 
 
     //重载坦克移动方法，便于我方坦克进行速度调整
-    public void moveUp(int v) {
+    public void moveUp(double v) {
         if (y > 0) {
             y -= speed*v;
         }
     }
 
-    public void moveRight(int v) {
+    public void moveRight(double v) {
         if (x + Attributes.OBJECT_SIZE < Attributes.WINDOW_WIDTH) {
             x += speed*v;
         }
     }
 
-    public void moveDown(int v) {
+    public void moveDown(double v) {
         if (y + Attributes.OBJECT_SIZE < Attributes.WINDOW_LENGTH) {
             y += speed*v;
         }
     }
 
-    public void moveLeft(int v) {
+    public void moveLeft(double v) {
         if (x > 0) {
             x -= speed*v;
         }
